[DDW-197] Moves layout themes into themes/simple
